[FEATURE] Add timeseries entity (#10)
* Adding support for abstract property

* Fixed new updates on Abstract Properties

* Fixed merge issues

* Adding support for abstract property

* Fixed new updates on Abstract Properties

* Fixed failing cypress tests

* Time Series Entity + Cypress tests

* Fixed small bugs

+++ core/libs/connection/src/lib/shape-connection-handler.ts
@@ -75,14 +75,15 @@ abstract class InheritanceConnector {
     protected mxGraphService: MxGraphService,
     protected mxGraphAttributeService: MxGraphAttributeService,
     protected languageSettingsService: LanguageSettingsService,
-    protected notificationsService?: NotificationsService
+    protected notificationsService: NotificationsService
   ) {}
 
   public connect(parentMetaModel: BaseMetaModelElement, childMetaModel: BaseMetaModelElement, parentCell: mxCell, childCell: mxCell) {
-    if ((childMetaModel as any).extendedElement === parentMetaModel) {
-      this.notificationsService.info(`${childMetaModel.name} already extends ${parentMetaModel.name}. Please remove the connection to create a new one.`, null, null, 5000);
+    if (typeof parentMetaModel['isPredefined'] === 'function' && parentMetaModel['isPredefined']()) {
+      this.notificationsService.warning("A predefined element can't have a child");
       return;
     }
+
     (parentMetaModel as any).extendedElement = childMetaModel;
     this.checkAndRemoveExtendElement(parentCell);
     this.mxGraphService.assignToParent(childCell, parentCell);
@@ -118,10 +119,11 @@ class EntityInheritanceConnector extends InheritanceConnector {
     protected mxGraphService: MxGraphService,
     protected mxGraphAttributeService: MxGraphAttributeService,
     protected languageSettingsService: LanguageSettingsService,
+    protected notificationsService: NotificationsService,
     protected propertyAbstractPropertyConnector?: PropertyAbstractPropertyConnectionHandler,
     protected entityPropertyConnector?: EntityPropertyConnectionHandler
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationsService);
   }
 
   connectWithAbstract(
@@ -171,9 +173,10 @@ class PropertyInheritanceConnector extends InheritanceConnector {
   constructor(
     protected mxGraphService: MxGraphService,
     protected mxGraphAttributeService: MxGraphAttributeService,
-    protected languageSettingsService: LanguageSettingsService
+    protected languageSettingsService: LanguageSettingsService,
+    protected notificationsService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationsService);
   }
 
   public connect(parentMetaModel: BaseMetaModelElement, childMetaModel: BaseMetaModelElement, parentCell: mxCell, childCell: mxCell) {
@@ -716,6 +719,11 @@ export class PropertyCharacteristicConnectionHandler implements ShapeMultiConnec
         outEdge.target.geometry.translate(ExpandedModelShape.expandedElementWidth, 0);
       }
 
+      const targetModel = MxGraphHelper.getModelElement(outEdge.target);
+      if (targetModel instanceof DefaultProperty || targetModel instanceof DefaultAbstractProperty) {
+        return;
+      }
+
       this.mxGraphService.removeCells([parent.removeEdge(outEdge, true)]);
     });
 
@@ -915,13 +923,27 @@ export class PropertyPropertyConnectionHandler
     protected languageSettingsService: LanguageSettingsService,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationService);
   }
 
   public connect(parentMetaModel: DefaultProperty, childMetaModel: DefaultProperty, parentCell: mxCell, childCell: mxCell) {
+    if (parentMetaModel.isPredefined()) {
+      this.notificationsService.warning("A predefined element can't have a child");
+      return;
+    }
+
     if (this.isRecursive(parentMetaModel, childCell)) {
       this.notificationService.warning('Recursive elements', 'Can not connect elements due to circular connection', null, 5000);
     } else {
+      if (childMetaModel.extendedElement) {
+        this.notificationService.warning(
+          'Illegal operation',
+          'Can not extend a Property which already extends another element',
+          null,
+          5000
+        );
+        return;
+      }
       super.connect(parentMetaModel, childMetaModel, parentCell, childCell);
     }
   }
@@ -940,7 +962,7 @@ export class PropertyAbstractPropertyConnectionHandler
     protected languageSettingsService: LanguageSettingsService,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationService);
   }
 
   public connect(parentMetaModel: DefaultProperty, childMetaModel: DefaultAbstractProperty, parentCell: mxCell, childCell: mxCell) {
@@ -965,7 +987,7 @@ export class AbstractPropertyAbstractPropertyConnectionHandler
     protected languageSettingsService: LanguageSettingsService,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationService);
   }
 
   public connect(parentMetaModel: DefaultAbstractProperty, childMetaModel: DefaultAbstractProperty, parentCell: mxCell, childCell: mxCell) {
@@ -1023,7 +1045,14 @@ export class EntityEntityConnectionHandler extends EntityInheritanceConnector im
     protected entityPropertyConnector: EntityPropertyConnectionHandler,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService, propertyAbstractPropertyConnector, entityPropertyConnector);
+    super(
+      mxGraphService,
+      mxGraphAttributeService,
+      languageSettingsService,
+      notificationService,
+      propertyAbstractPropertyConnector,
+      entityPropertyConnector
+    );
   }
 
   public connect(parentMetaModel: DefaultEntity, childMetaModel: DefaultEntity, parentCell: mxCell, childCell: mxCell) {
@@ -1050,7 +1079,7 @@ export class AbstractEntityAbstractEntityConnectionHandler
     protected languageSettingsService: LanguageSettingsService,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService);
+    super(mxGraphService, mxGraphAttributeService, languageSettingsService, notificationService);
   }
 
   public connect(parentMetaModel: DefaultAbstractEntity, childMetaModel: DefaultAbstractEntity, parentCell: mxCell, childCell: mxCell) {
@@ -1077,7 +1106,14 @@ export class EntityAbstractEntityConnectionHandler
     protected entityPropertyConnector: EntityPropertyConnectionHandler,
     private notificationService: NotificationsService
   ) {
-    super(mxGraphService, mxGraphAttributeService, languageSettingsService, propertyAbstractPropertyConnector, entityPropertyConnector);
+    super(
+      mxGraphService,
+      mxGraphAttributeService,
+      languageSettingsService,
+      notificationService,
+      propertyAbstractPropertyConnector,
+      entityPropertyConnector
+    );
   }
 
   public connect(parentMetaModel: DefaultEntity, childMetaModel: DefaultAbstractEntity, parent: mxCell, child: mxCell) {
